removed all commented code, underlined active page link, moved the missions.js slice file to a /missions folder, changed the key features section in the readme.md file, adjusted the color of the reserved rockets tag
