+++ src/App.css
@@ -81,9 +81,9 @@
 }
 
 .active-link {
-  background-color: #0084ff;
-  border-radius: 5px;
-  color: white !important;
+  border-bottom: 2px solid #0084ff;
+  padding: 1px;
+  width: 75%;
 }
 
 .profile {
