+++ src/components/styles/rockets.module.css
@@ -68,7 +68,7 @@
 }
 
 .reservedBanner {
-  background: #1da5ff;
+  background: #19acb1;
   padding: 0.15rem 0.5rem;
   margin-right: 0.5rem;
   font-size: 0.9rem;
